Implement 64-bit node uuid

+++ src/gnb_conf_type.h
@@ -62,7 +62,7 @@ typedef struct _gnb_conf_t {
 	uint8_t public_index_service;
 	uint8_t lite_mode;
 
-	uint32_t local_uuid;
+	uint64_t local_uuid;
 
 	char binary_dir[PATH_MAX];
